chore(ACI): Metric alert detector nits
