+++ src/sentry/incidents/metric_alert_detector.py
@@ -1,5 +1,5 @@
 from datetime import timedelta
-from typing import Any, TypedDict
+from typing import Any
 
 from rest_framework import serializers
 
@@ -13,26 +13,11 @@ from sentry.workflow_engine.endpoints.validators.base import (
 )
 from sentry.workflow_engine.models import DataConditionGroup, DataSource, Detector
 from sentry.workflow_engine.models.data_condition import Condition, DataCondition
-from sentry.workflow_engine.types import DetectorPriorityLevel
-
-
-class DataConditionType(TypedDict):
-    id: int | None
-    comparison: int
-    type: Condition
-    condition_result: DetectorPriorityLevel
-    condition_group_id: int
-
-
-class DataSourceType(TypedDict):
-    query_type: int
-    dataset: str
-    query: str
-    aggregate: str
-    time_window: float
-    resolution: float
-    environment: str
-    event_types: list[SnubaQueryEventType]
+from sentry.workflow_engine.types import (
+    DataConditionType,
+    DetectorPriorityLevel,
+    SnubaQueryDataSourceType,
+)
 
 
 class MetricAlertComparisonConditionValidator(NumericComparisonConditionValidator):
@@ -100,7 +85,7 @@ class MetricAlertsDetectorValidator(BaseDetectorTypeValidator):
                 )
         return data_condition_group
 
-    def update_data_source(self, instance: Detector, data_source: DataSourceType):
+    def update_data_source(self, instance: Detector, data_source: SnubaQueryDataSourceType):
         try:
             source_instance = DataSource.objects.get(detector=instance)
         except DataSource.DoesNotExist:
@@ -138,7 +123,7 @@ class MetricAlertsDetectorValidator(BaseDetectorTypeValidator):
         if data_conditions:
             self.update_data_conditions(instance, data_conditions)
 
-        data_source: DataSourceType = validated_data.pop("data_source")
+        data_source: SnubaQueryDataSourceType = validated_data.pop("data_source")
         if data_source:
             self.update_data_source(instance, data_source)
