+++ src/sentry/workflow_engine/types.py
@@ -9,7 +9,9 @@ if TYPE_CHECKING:
     from sentry.deletions.base import ModelRelation
     from sentry.eventstore.models import GroupEvent
     from sentry.eventstream.base import GroupState
+    from sentry.snuba.models import SnubaQueryEventType
     from sentry.workflow_engine.models import Action, Detector, Workflow
+    from sentry.workflow_engine.models.data_condition import Condition
 
 T = TypeVar("T")
 
@@ -71,3 +73,22 @@ class DataConditionHandler(Generic[T]):
     @staticmethod
     def evaluate_value(value: T, comparison: Any) -> DataConditionResult:
         raise NotImplementedError
+
+
+class DataConditionType(TypedDict):
+    id: int | None
+    comparison: int
+    type: Condition
+    condition_result: DetectorPriorityLevel
+    condition_group_id: int
+
+
+class SnubaQueryDataSourceType(TypedDict):
+    query_type: int
+    dataset: str
+    query: str
+    aggregate: str
+    time_window: float
+    resolution: float
+    environment: str
+    event_types: list[SnubaQueryEventType]
